p2p-governor: test asynchronous demotiions
This is only enabled in the re-enabled `prop_governor_nolivelock`.
Running the test throws an exception (failed assertion), this is a bug
fixed in the next commit.

If a peer during asynchronous demotion is demoted to 'ColdPeer', we
interpret this as an error and throw an exception.  This corresponds to
a real scenario.  This is important because otherwise the governor will
return and finish the transition without running its error handlers
which bring the governor to the right state.

The above corresponds to the property of 'PeerStateActions': on failures
the peer state changes to 'PeerCold'.