Weigthed random demotion policies from KnownPeerInfo
Make KnownPeerInfo visible to PeerSelectionPolicy so that error count
and the tepid flag (was Hot, but got demoted due to poor performance)
can be used to influence peer ranking.

Peers with knownPeerTepid set are twice as likely to be picked for
warm to cold demotion.
The higher the knownPeerFailCount the higher the chance that a peer will
be picked to be forgotten.